fix Base64ManagerImpl and PeraFixture for iOS in KMP
