+++ app/src/main/java/com/algorand/android/modules/algosdk/domain/usecase/BuildKeyRegOnlineTransactionImpl.kt
@@ -12,6 +12,7 @@
 
 package com.algorand.android.modules.algosdk.domain.usecase
 
+import android.util.Log
 import com.algorand.algosdk.builder.transaction.KeyRegistrationTransactionBuilder
 import com.algorand.android.modules.algosdk.domain.mapper.TransactionParametersResponseMapper
 import com.algorand.android.modules.algosdk.domain.model.OnlineKeyRegTransactionPayload
@@ -24,7 +25,8 @@ internal class BuildKeyRegOnlineTransactionImpl @Inject constructor(
     override fun invoke(params: OnlineKeyRegTransactionPayload): ByteArray? {
         return try {
             createTransaction(params)
-        } catch (exception: Exception) {
+        } catch (e: Exception) {
+            Log.e("BuildKeyRegOnlineTransactionImpl", e.toString())
             null
         }
     }
